# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: 3GPP Server API
  description: |
    A JSON-to-NGAP/NAS translation server for conformance testing, fuzzing,
    and security assessment of 5G core networks.

    The server manages gNB and UE contexts, translates JSON requests into
    NGAP/NAS messages over SCTP, and returns decoded responses as JSON.

    ## Workflow

    1. **Create a gNB** — dials SCTP to the AMF, performs NG Setup.
    2. **Create a UE** — provisions credentials (K, OPc), computes SUCI.
    3. **Send NGAP messages** — builds NAS inside NGAP, sends over SCTP,
       returns the decoded AMF response.

    ## IE Control

    The JSON schema mirrors NGAP/NAS IEs 1:1. Every optional IE can be
    individually set, overridden, or omitted — enabling both correct
    procedures and deliberate deviations for fuzz testing.

    ## 3GPP References

    - NGAP: TS 38.413
    - NAS 5GS: TS 24.501
    - 5G Security: TS 33.501
  version: "0.2.0"
  license:
    name: BUSL-1.1
    identifier: BUSL-1.1
paths:
...
  /gnb/{gnb_id}/ngap:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
    post:
      operationId: sendGnBNGAP
      summary: Send a gNB-level (non-UE-associated) NGAP message
      description: |
        Sends a non-UE-associated NGAP message on the gNB's existing N2
        association and returns the decoded AMF response.

        ## Supported message types

        - `ng_reset` — Sends an NG RESET (TS 38.413 §8.7.4). With no
          `reset_ue_ids` it resets the whole NG interface; with `reset_ue_ids`
          it resets only those UEs' associations (partOfNG-Interface). The AMF
          answers with NG RESET ACKNOWLEDGE.
        - `handover_request_acknowledge` — Target-gNB HANDOVER REQUEST
          ACKNOWLEDGE (TS 38.413 §8.4.2), admitting the listed `pdu_sessions`
          with their downlink GTP tunnels. Send-only.
        - `handover_notify` — Target-gNB HANDOVER NOTIFY once the UE has
          arrived (TS 38.413 §8.4.3). Send-only.
        - `path_switch_request` — NG-RAN-node PATH SWITCH REQUEST (TS 38.413
          §8.4.4), switching the downlink path of the existing UE context
          identified by `amf_ue_ngap_id` (its Source AMF UE NGAP ID) toward the
          GTP tunnels in `pdu_sessions`. `ran_ue_ngap_id` is the new RAN UE NGAP
          ID. With `wait_for` the call blocks for the AMF's PATH SWITCH REQUEST
          ACKNOWLEDGE or PATH SWITCH REQUEST FAILURE.

        ## Raw escape hatch

        `raw_ngap_pdu` (hex) is the universal primitive: when set,
        `message_type` is ignored and the bytes are written verbatim onto the
        N2 association with no encoding or validation — letting a test craft any
        NGAP PDU at all (well-formed, malformed, out-of-sequence, or garbage) to
        probe how the core reacts. With `wait_for` the call blocks for one of
        those downlink message types and returns it; otherwise it is
        fire-and-forget.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message_type:
                  type: string
                  enum: [ng_reset, handover_request_acknowledge, handover_failure, handover_notify, path_switch_request]
                raw_ngap_pdu:
                  type: string
                  description: "Hex NGAP PDU sent verbatim (overrides message_type). Craft any message, valid or not."
                wait_for:
                  type: array
                  items:
                    type: string
                  description: "Downlink NGAP message types to block for (raw_ngap_pdu, path_switch_request). Empty = fire-and-forget."
                timeout_ms:
                  type: integer
                  description: "With wait_for: how long to wait. Default 5000."
                reset_ue_ids:
                  type: array
                  items:
                    type: string
                  description: "UE IDs whose NG associations to reset (NG Reset). Empty = reset all."
                amf_ue_ngap_id:
                  type: integer
                  description: "AMF UE NGAP ID identifying the UE association (Handover; Path Switch Request Source AMF UE NGAP ID)."
                cause:
                  type: integer
                  description: "Radio-network Cause for handover_failure (TS 38.413 §9.3.1.2). Defaults to ho-failure-in-target-5GC."
                ran_ue_ngap_id:
                  type: integer
                  description: "RAN UE NGAP ID (Handover target-assigned; Path Switch Request new NG-RAN-node ID)."
                pdu_sessions:
                  type: array
                  description: "PDU sessions with their downlink GTP tunnels (Handover Request Acknowledge admitted list; Path Switch Request to-be-switched list)."
                  items:
                    type: object
                    properties:
                      id: { type: integer }
                      dl_teid: { type: integer }
                      dl_ip: { type: string }
                      raw_transfer:
                        type: string
                        description: "Hex transfer that replaces the built one verbatim, for crafting malformed transfers."
                ue_security_capabilities:
                  type: object
                  description: "Path Switch Request: UE security capability bitmaps (TS 38.413 §9.3.1.86). Each is 4 hex digits (16 bits). Defaults: NR NEA1-3/NIA1-3, E-UTRA none."
                  properties:
                    nr_encryption: { type: string }
                    nr_integrity: { type: string }
                    eutra_encryption: { type: string }
                    eutra_integrity: { type: string }
                omit_ies:
                  type: array
                  items:
                    type: integer
                  description: "Path Switch Request: protocol IE ids to drop from the built message, for sending a structurally-incomplete message (e.g. missing a mandatory IE). For fully arbitrary bytes use raw_ngap_pdu."
                failed_pdu_sessions:
                  type: array
                  items:
                    type: integer
                  description: "PDU sessions reported as failed-to-setup (Handover Request Acknowledge; Path Switch Request)."
      responses:
        "200":
          description: Decoded AMF response.
        "404":
          description: gNB not found.
        "504":
          description: No response from the AMF.
  /gnb/{gnb_id}/await:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
    post:
      operationId: awaitGnBMessage
      summary: Await an unsolicited downlink NGAP message
      description: |
        Blocks until one of the given NGAP message types arrives on the gNB's
        N2 association, then returns the decoded message. Used for messages the
        core sends without a directly-triggering uplink — e.g. HANDOVER REQUEST
        on a target gNB, or HANDOVER COMMAND / UE CONTEXT RELEASE COMMAND on a
        source gNB during N2 handover.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message_types]
              properties:
                message_types:
                  type: array
                  items:
                    type: string
                  description: "NGAP message types to wait for (first match wins)."
                timeout_ms:
                  type: integer
                  description: "How long to wait. Default 5000."
      responses:
        "200":
          description: Decoded NGAP message.
        "404":
          description: gNB not found.
        "504":
          description: No matching message arrived before the timeout.
...
  /gnb/{gnb_id}/ue/{ue_id}/await:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
      - $ref: "#/components/parameters/ue_id"
    post:
      operationId: awaitUEMessage
      summary: Await an unsolicited downlink NGAP message for a specific UE
      description: |
        Like the gNB-level await, but only returns a downlink PDU addressed to
        this UE (matched by its RAN/AMF UE NGAP ID). This lets many UEs share one
        gNB association without consuming each other's downlink — needed for
        concurrency and fast-sequence scenarios. A PDU that carries no UE NGAP ID
        (e.g. an association-level Error Indication) matches any waiter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message_types]
              properties:
                message_types:
                  type: array
                  items:
                    type: string
                  description: "NGAP message types to wait for (first match wins)."
                timeout_ms:
                  type: integer
                  description: "How long to wait. Default 5000."
      responses:
        "200":
          description: Decoded NGAP message addressed to the UE.
        "404":
          description: gNB or UE not found.
        "504":
          description: No matching message arrived before the timeout.
...
  /gnb/{gnb_id}/ue/{ue_id}/ngap:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
      - $ref: "#/components/parameters/ue_id"
    post:
      operationId: sendNGAP
      summary: Send NGAP message
      description: |
        Builds an NGAP message using stored gNB/UE state, sends it over SCTP,
        waits for the AMF response, and returns the decoded NGAP + NAS.

        The request body controls which NAS message to send and which optional
        IEs to include. Any field present overrides the stored default.
        Omitted fields use sensible defaults from the UE context.

        **Does NOT auto-respond** — the caller controls the entire flow.

        ## Supported message types

        - `registration_request` — Sends RegistrationRequest inside InitialUEMessage.
        - `authentication_response` — Computes 5G-AKA RES* from stored RAND/AUTN, sends AuthenticationResponse in UplinkNASTransport.
        - `security_mode_complete` — Derives NAS keys, sends SecurityModeComplete (with integrity+ciphering) in UplinkNASTransport.
        - `registration_complete` — Sends InitialContextSetupResponse + RegistrationComplete (with integrity+ciphering) in UplinkNASTransport.

        ## NAS IE overrides (TS 24.501 §8.2.6)

        For `registration_request`, every optional IE in the RegistrationRequest
        can be individually set via the corresponding JSON field. Hex-encoded
        strings are used for binary IEs. This enables both valid and deliberately
        malformed messages for fuzz testing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNGAPRequest"
      responses:
        "200":
          description: NGAP message sent and response received.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendNGAPResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
components:
  parameters:
    gnb_id:
      name: gnb_id
      in: path
      required: true
      schema:
        type: string
    ue_id:
      name: ue_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: SCTP or NGAP error communicating with the AMF.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    GatewayTimeout:
      description: Timed out waiting for AMF response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SendNGAPRequest:
      type: object
      required: [message_type]
      properties:
        message_type:
          type: string
          description: NAS message to send.
          enum: [registration_request, authentication_response, security_mode_complete, registration_complete, pdu_session_establishment_request, deregistration_request, ue_context_release_request, service_request, identity_response, pdu_session_release_request, pdu_session_release_complete, pdu_session_modification_request, pdu_session_modification_complete, pdu_session_modification_command_reject, status_5gsm, authentication_failure, security_mode_reject, handover_required, handover_cancel, ran_status_transfer, inject_nas, ue_capability_info, error_indication, initial_context_setup_failure]
        raw_nas_pdu:
          type: string
          description: |
            Raw NAS PDU override (hex). When set, all NAS building and security
            encoding is skipped — these bytes are sent directly in the NGAP wrapper.
            Use for fuzz testing with arbitrary/malformed NAS content.
        rrc_establishment_cause:
          type: integer
          description: Override RRC establishment cause in InitialUEMessage (registration_request only).
        ue_context_request:
          type: integer
          description: Override UE Context Request in InitialUEMessage. Set to -1 to omit entirely.
        amf_ue_ngap_id_override:
          type: integer
          description: Override the AMF UE NGAP ID in UE-associated NGAP messages (e.g. UplinkNASTransport, handover_required). Use a stale/unknown value to exercise TS 38.413 §10.6.
        ran_ue_ngap_id_override:
          type: integer
          description: Override RAN UE NGAP ID in the NGAP wrapper.
        res_star_override:
          type: string
          description: Override RES* in authentication_response (hex, 16 bytes). Skips 5G-AKA computation.
        corrupt_mac:
          type: boolean
          description: On a security-protected message (security_mode_complete/service_request/mobility registration_request), flip a NAS-MAC byte so the AMF's integrity check fails and it discards the message (TS 24.501 §4.4.4.3).
        nas_count:
          type: integer
          description: Force the uplink NAS COUNT written into a built message (e.g. a stale value) for replay/robustness tests; the UE's real counter still advances.
        replay_last:
          type: boolean
          description: On inject_nas, resend this UE's last secured uplink NAS PDU verbatim (NAS replay protection test, TS 24.501 §4.4.3.2).
        status_transfer_drbs:
          type: array
          description: |
            On ran_status_transfer, the DRBs whose PDCP status the source NG-RAN
            node reports (TS 38.413 §8.4.6.2). Defaults to one DRB with zeroed COUNTs.
          items:
            type: object
            properties:
              drb_id: { type: integer }
              ul_pdcp_sn: { type: integer }
              ul_hfn: { type: integer }
              dl_pdcp_sn: { type: integer }
              dl_hfn: { type: integer }
        timeout_ms:
          type: integer
          description: Bounds how long a send waits for a downlink. Default 5000.
        ue_radio_capability:
          type: string
          description: On ue_capability_info, the hex UE radio capability the gNB reports; the AMF replays it in a later Initial Context Setup Request.
        release_cause:
          type: integer
          description: "Radio-network Cause value for ue_context_release_request (TS 38.413 §9.3.1.2). Default 20 (user-inactivity)."
        service_type:
          type: integer
          description: "Service type for service_request (TS 24.501 §9.11.3.50). Default 1 (data). 0=signalling, 2=mobile-terminated."
        target_gnb_id:
          type: string
          description: "Target gNB identity (hex) for handover_required (TS 38.413 §8.4.1). The resulting Handover Command arrives asynchronously — await it on the source gNB."
        pdu_session_ids:
          type: array
          items:
            type: integer
          description: "PDU sessions to hand over in handover_required. Defaults to the UE's session; set to non-existent IDs to exercise abnormal cases."
        pdu_session_id:
          type: integer
          description: "Which PDU session a pdu_session_establishment_request establishes, so a UE can hold more than one."
        request_type:
          type: integer
          description: "Request Type IE of the UL NAS Transport carrying a 5GSM message (TS 24.501 §9.11.3.47): 1=initial request, 2=existing PDU session, 3=initial emergency, 4=existing emergency, 5=modification request, 7=reserved. When unset, no Request Type IE is sent. Applies to the pdu_session_modification_request / pdu_session_release_request endpoints."
        pti:
          type: integer
          description: "Procedure Transaction Identity of a 5GSM message (TS 24.501 §9.6, §7.3.1): 0=unassigned, 1-254=assigned, 255=reserved. When unset, an assigned value (1) is used. Applies to the pdu_session_establishment_request / pdu_session_modification_request / pdu_session_release_request / pdu_session_release_complete / pdu_session_modification_complete / pdu_session_modification_command_reject / status_5gsm endpoints."
        always_on_requested:
          type: boolean
          description: "Add the Always-on PDU session requested IE to a pdu_session_establishment_request (TS 24.501 §8.3.1.7). The network must answer with an Always-on PDU session indication in the Establishment Accept (§6.4.1)."
        cause_5gsm:
          type: integer
          description: "5GSM cause for a UE-originated reject/status message (TS 24.501 §9.11.4.2). Applies to the pdu_session_modification_command_reject / status_5gsm endpoints. Defaults to #111 protocol error, unspecified."
        handover_required_cause:
          type: integer
          description: "Radio-network Cause the source NG-RAN node reports on handover_required (TS 38.413 §9.3.1.2). Defaults to handover-desirable-for-radio-reason."
        handover_cancel_cause:
          type: integer
          description: "Radio-network Cause for handover_cancel (TS 38.413 §9.3.1.2). Defaults to handover-cancelled."
        existing_connection:
          type: boolean
          description: "Send a registration_request over the existing UE-associated connection (Uplink NAS Transport) instead of an Initial UE Message — e.g. a Mobility Registration Update after an N2 handover."
        switch_off:
          type: integer
          description: "Deregistration switch-off flag (TS 24.501 §9.11.3.20). Default 1 (switch off, no accept). 0 = normal de-registration (AMF replies with Deregistration Accept)."
        registration_type:
          type: integer
          description: "5GS registration type: 1=initial, 2=mobility, 3=periodic, 4=emergency."
        ng_ksi:
          type: integer
          description: NAS Key Set Identifier override.
        follow_on_request:
          type: integer
          description: "FOR bit (0 or 1). Default 1."
        mobile_identity_override:
          type: string
          description: Override mobile identity (hex). Replaces SUCI/GUTI.
        ue_security_capability:
          type: string
          description: Override UE security capability (hex, 2-8 bytes).
        requested_nssai:
          type: array
          items:
            type: object
            properties:
              sst:
                type: integer
              sd:
                type: string
        capability_5gmm:
          type: string
          description: 5GMM capability (hex).
        non_current_native_nas_ksi:
          type: integer
        last_visited_registered_tai:
          type: string
          description: Last visited TAI (hex, 6 bytes).
        s1_ue_network_capability:
          type: string
          description: S1 UE network capability (hex).
        uplink_data_status:
          type: string
          description: Uplink data status (hex, 2+ bytes).
        pdu_session_status:
          type: string
          description: PDU session status (hex, 2+ bytes).
        mico_indication:
          type: integer
          description: "MICO indication (RAAI bit)."
        ue_status:
          type: integer
          description: "UE status (bit 0: S1 mode, bit 1: N1 mode)."
        additional_guti:
          type: string
          description: Additional GUTI (hex).
        allowed_pdu_session_status:
          type: string
          description: Allowed PDU session status (hex).
        ues_usage_setting:
          type: integer
        requested_drx_parameters:
          type: integer
        eps_nas_message_container:
          type: string
          description: EPS NAS message container (hex).
        ladn_indication:
          type: string
          description: LADN indication (hex).
        payload_container:
          type: string
          description: Payload container (hex).
        network_slicing_indication:
          type: integer
          description: "Network slicing indication (bit 0: NSSCI, bit 1: DCNI)."
        update_type_5gs:
          type: string
          description: 5GS update type (hex).
        nas_message_container:
          type: string
          description: NAS message container (hex).
        eps_bearer_context_status:
          type: string
          description: EPS bearer context status (hex, 2 bytes).
    SendNGAPResponse:
      type: object
      properties:
        ngap:
          $ref: "#/components/schemas/NGAPResponse"
        nas:
          $ref: "#/components/schemas/NASResponse"
    NGAPResponse:
      type: object
      properties:
        pdu_type:
          type: string
          enum: [initiating_message, successful_outcome, unsuccessful_outcome]
        message_type:
          type: string
          description: "Decoded message name (e.g. NGSetupResponse, DownlinkNASTransport)."
        raw_hex:
          type: string
          description: Full NGAP PDU as hex.
        ies:
          type: array
          items:
            type: object
            description: Decoded IE with typed fields.
...